requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
PyYAML==6.0.2
python-dotenv==1.0.1
fpdf2==2.7.9
openai==1.42.0
praw==7.8.1
openai>=1.30.0,<2
httpx==0.27.2
praw>=7.8.1